Changing Audit Ids to BIGINT

+++ shoppingcart-api/src/sql/trigger/trAddress.trigger.sql
@@ -13,7 +13,7 @@ CREATE TRIGGER trAddress
 	SET NOCOUNT ON
 
 	DECLARE 
-		@AuditLogTransactionId	int,
+		@AuditLogTransactionId	bigint,
 		@Inserted	    		int = 0,
  		@ROWS_COUNT				int
